+++ python/aistore/sdk/response_handler.py
@@ -44,7 +44,7 @@ class ResponseHandler(ABC):
             # Raise specific error type but keep full HTTPError with stack trace
             raise self.parse_error(r) from http_err
         # If not an expected status code, raise general error
-        raise self.exc_class(r.status_code, r.text, r.request.url or "")
+        raise self.exc_class(r.status_code, r.text, r.request.url or "", r.request)
 
     @abstractmethod
     def parse_error(self, r: requests.Response) -> APIRequestError:
@@ -82,7 +82,7 @@ class AISResponseHandler(ResponseHandler):
             ErrETLNotFound: If the error message indicates a missing ETL.
         """
         status, message, req_url = r.status_code, r.text, r.request.url
-        prov, bck, has_obj = extract_and_parse_url(message) or (None, None, None)
+        prov, bck, obj = extract_and_parse_url(message) or (None, None, None)
 
         if prov is not None:
             try:
@@ -97,23 +97,22 @@ class AISResponseHandler(ResponseHandler):
 
         exc = self.exc_class
         if status == 404:
-            exc = self._parse_404(is_etl_err, is_remote_bck, prov, has_obj) or exc
+            exc = self._parse_404(is_etl_err, is_remote_bck, prov, obj) or exc
         elif status == 409:
-            exc = self._parse_409(is_etl_err, is_get_request, prov, has_obj) or exc
-
-        return exc(status, message, req_url or "")
+            exc = self._parse_409(is_etl_err, is_get_request, prov, obj) or exc
+        return exc(status, message, req_url or "", r.request)
 
     @staticmethod
     def _parse_404(
         is_etl_err: bool,
         is_remote_bck: Optional[bool] = None,
         provider: Optional[Provider] = None,
-        has_obj: Optional[bool] = None,
+        obj: Optional[bool] = None,
     ) -> Optional[Type[AISError]]:
         if is_etl_err:
             return ErrETLNotFound
         if provider:
-            if has_obj:
+            if obj is not None:
                 return ErrObjNotFound
             if is_remote_bck:
                 return ErrRemoteBckNotFound
@@ -125,12 +124,12 @@ class AISResponseHandler(ResponseHandler):
         is_etl_err: bool,
         is_get_request: bool,
         provider: Optional[Provider] = None,
-        has_obj: Optional[bool] = None,
+        obj: Optional[bool] = None,
     ) -> Optional[Type[AISError]]:
         if is_get_request:
             return ErrGETConflict
         if is_etl_err:
             return ErrETLAlreadyExists
-        if provider and not has_obj:
+        if provider and obj is None:
             return ErrBckAlreadyExists
         return None