+++ python/tests/unit/sdk/test_ais_response_handler.py
@@ -10,7 +10,7 @@ from aistore.sdk.errors import (
     ErrETLNotFound,
 )
 from aistore.sdk.response_handler import AISResponseHandler
-from tests.utils import cases, create_api_error_response
+from tests.utils import cases, handler_parse_and_assert
 
 
 # pylint: disable=unused-variable
@@ -35,13 +35,4 @@ class TestAISResponseHandler(unittest.TestCase):
         ("etl job test-etl-job does not exist", ErrETLNotFound, 404),
     )
     def test_parse_ais_error(self, test_case):
-        err_msg, expected_err, err_status = test_case
-        test_url = "http://test-url"
-        response = create_api_error_response(test_url, err_status, err_msg)
-
-        err = AISResponseHandler().parse_error(response)
-        self.assertIsInstance(err, AISError)
-        self.assertIsInstance(err, expected_err)
-        self.assertEqual(err_status, err.status_code)
-        self.assertEqual(err_msg, err.message)
-        self.assertEqual(test_url, err.req_url)
+        handler_parse_and_assert(self, AISResponseHandler(), AISError, test_case)