+++ python/tests/unit/sdk/test_bucket.py
@@ -2,6 +2,7 @@ import unittest
 from unittest.mock import Mock, call, patch, MagicMock, mock_open
 from urllib.parse import quote
 
+from requests import PreparedRequest
 from requests.structures import CaseInsensitiveDict
 
 from aistore.sdk.ais_source import AISSource
@@ -149,7 +150,9 @@ class TestBucket(unittest.TestCase):
         self._assert_bucket_created(res)
 
     def test_create_already_exists(self):
-        already_exists_err = ErrBckAlreadyExists(400, "message", "bck_create_url")
+        already_exists_err = ErrBckAlreadyExists(
+            400, "message", "bck_create_url", Mock(PreparedRequest)
+        )
         self.mock_client.request.side_effect = already_exists_err
         with self.assertRaises(ErrBckAlreadyExists):
             self.ais_bck.create()
@@ -195,7 +198,7 @@ class TestBucket(unittest.TestCase):
 
     def test_delete_missing(self):
         self.mock_client.request.side_effect = ErrBckNotFound(
-            400, "not found", "bck_delete_url"
+            400, "not found", "bck_delete_url", Mock(PreparedRequest)
         )
         with self.assertRaises(ErrBckNotFound):
             Bucket(client=self.mock_client, name="missing-bucket").delete()