+++ python/tests/unit/sdk/test_session_manager.py
@@ -7,7 +7,6 @@ import urllib3
 
 from aistore.sdk.const import AIS_CLIENT_CA, AIS_CLIENT_CRT, AIS_CLIENT_KEY
 from aistore.sdk.session_manager import SessionManager
-from aistore.sdk.retry_config import RetryConfig
 
 from tests.utils import cases
 
@@ -17,12 +16,8 @@ class TestSessionManager(unittest.TestCase):  # pylint: disable=unused-variable
         self.endpoint = "https://aistore-endpoint"
         self.mock_session = Mock()
 
-    @patch("aistore.sdk.retry_config.RetryConfig.default")
-    def test_init_default(self, mock_retry_config):
-        retry_config = RetryConfig.default()
-        mock_retry_config.return_value = retry_config
+    def test_init_default(self):
         session_manager = SessionManager()
-        self.assertEqual(retry_config.http_retry, session_manager.retry)
         self.assertIsNone(session_manager.ca_cert)
         self.assertFalse(session_manager.skip_verify)